updated version 2.45.0 return module place order module done

+++ src/com/devstack/pos/repository/OrderDetailRepository.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
     
     @Query("SELECT o FROM OrderDetail o WHERE o.customerId = :customerId ORDER BY o.issuedDate DESC")
     List<OrderDetail> findByCustomerId(@Param("customerId") Long customerId);
